chore(math): update comments to contain montgomery form && fix minors
Fot the 31-bit prime field, the montgomery form is `a * R`, whre R is 2³².
So value `a` can be ranged in {0, ..., P}, whre P is modulus. The
comments are updated to handle this case, too.

Additionally, this commit also contiain minor fixes, such as chaning `P`
to `p` and removing comma and period.